Fix issue of even-numbered rounds not working

+++ scratch/sim/MyApp.h
@@ -29,15 +29,15 @@ private:
     Ptr<Socket> m_socket{nullptr};
     Address m_peer;
     // uint32_t m_packetSize{0}; // Removing m_packetSize as it's not used for scheduling or sending payload size calculation
-    uint32_t m_nPackets{0};
+    uint64_t m_nPackets{0};
     DataRate m_dataRate;
     EventId m_sendEvent;
     bool m_running{false};
-    uint32_t m_packetsSent{0};
+    uint64_t m_packetsSent{0};
     Time m_startTime; // Stores the app start time
     uint8_t* m_data; // Pointer to the normal data buffer (owned by caller)
     uint8_t* m_data_fin; // Pointer to the final packet data buffer (owned by caller)
-    uint32_t m_writeSize; // Actual size of payload for each packet
+    uint64_t m_writeSize; // Actual size of payload for each packet
 
 };
